+++ src/Plisky.Nuke.Fusion/Versonify/VersionifySettings.cs
@@ -6,6 +6,11 @@ namespace Plisky.Nuke.Fusion;
 public class VersonifySettings : ToolOptions {
 
     /// <summary>
+    /// Corresponds to -z on the versonify tool.
+    /// </summary>
+    public bool AlwaysReturnZero { get; set; } = false;
+
+    /// <summary>
     /// Corresponds to -O in the versionify tool.  Can be env, con, file, np, npo, azdo
     /// </summary>
     public string OutputStyle { get; set; } = string.Empty;
@@ -72,7 +77,7 @@ public class VersonifySettings : ToolOptions {
 
     public bool DryRun { get; set; }
 
-    public string Framework { get; set; }
+    public string? Framework { get; set; }
 
     public string GetPath() {
         string? fw = string.IsNullOrEmpty(Framework) ? null : Framework;
@@ -82,7 +87,10 @@ public class VersonifySettings : ToolOptions {
           framework: fw);
     }
 
-    public ArgumentStringHandler GetArgsString() {
+    public ArgumentStringHandler GetArgsAsString200() {
+        // Compatible with V1.0.2 of Versonify.  This is the current version as the time of writing this code.
+        // Indicated by 200 level return code from versonify.
+
         var result = new ArgumentStringHandler(0, 0, out _);
         result.AppendLiteral(Command);
         result.AppendLiteral($" -v={VersionPersistanceValue}");
@@ -94,6 +102,53 @@ public class VersonifySettings : ToolOptions {
 
         if (!string.IsNullOrEmpty(MultiMatchFile)) {
             result.AppendLiteral($" -m={MultiMatchFile}");
+        }
+
+        if (Debug) {
+            result.AppendLiteral(" -Debug");
+        }
+
+        if (DryRun) {
+            result.AppendLiteral(" -DryRun");
+        }
+
+        if (PerformIncrement) {
+            result.AppendLiteral(" -Increment");
+        }
+
+        if (AlwaysReturnZero) {
+            result.AppendLiteral(" -z");
+        }
+
+        if (!string.IsNullOrEmpty(TraceConfiguration)) {
+            result.AppendLiteral($" -Trace={TraceConfiguration}");
+        }
+
+        if (OutputStyle != string.Empty) {
+            if (!OutputStyle.EndsWith("-nf")) {
+                OutputStyle += "-nf";
+            }
+            result.AppendLiteral($" -O={OutputStyle}");
+        } else {
+            result.AppendLiteral(" -O=con-nf");
+        }
+        return result;
+    }
+
+    public ArgumentStringHandler GetArgsString() {
+        // This is the first implementaiton, compatible with versions below 1.0.2 - no return code from versonify.
+
+        var result = new ArgumentStringHandler(0, 0, out _);
+        result.AppendLiteral(Command);
+        result.AppendLiteral($" -vs={VersionPersistanceValue}");
+        result.AppendLiteral($" -Root={Root}");
+
+        if (!string.IsNullOrEmpty(QuickValue)) {
+            result.AppendLiteral($" -Q={QuickValue}");
+        }
+
+        if (!string.IsNullOrEmpty(MultiMatchFile)) {
+            result.AppendLiteral($" -mm={MultiMatchFile}");
         }
 
         if (Debug) {