+++ src/_Dependencies/Packaging/changelog.md
@@ -1,4 +1,9 @@
-Plisky.Nuke.Fusion Change Log. 
+﻿Plisky.Nuke.Fusion Change Log. 
+ 
+⬆️ V XXX-VERSION-XXX - Austen Compatibility Release.
+  - ✅ Feature - Updated Versonify support to be compatible with Austen.
+  
+  
 
 V0.3 
   * Fix to support Nuke 9.04